REF: Keep subs property with _subs attribute definition.
